move tests out of templates folder
