+++ docker-compose.yml
@@ -10,6 +10,7 @@ services:
     restart: unless-stopped
   pgadmin:
     # use hostname host.docker.internal to connect instead of localhost
+    # and use db user "knex"
     container_name: pgadmin_container
     image: dpage/pgadmin4
     ports:
