--- tests/templates/put.test.js
@@ -1,5 +1,0 @@
-describe.skip('put', () => {
-  test('something', () => {
-    //
-  })
-})
