+++ tests/update.test.js
@@ -1,0 +1,7 @@
+require('./utils/trxify-tests')
+
+describe.skip('put', () => {
+  test('prevent setting an invalid field', async () => {
+    //
+  })
+})
